requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2
django-cors-headers==3.7.0
django-redis==4.12.1
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
Markdown==3.3.4
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.9
Pillow==8.2.0
PyJWT==1.7.1
pyparsing==2.4.7
pytz==2021.1
redis==3.5.3
requests==2.25.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.4
